requirements.txt
gpiozero
websockets
PyYAML

